requirements.txt
pandas==1.3.5
azure-identity==1.11.0
azure-storage-file-datalake==12.9.1
azure-storage-blob==12.14.1
pyarrow==10.0.0
xlsxwriter==3.0.3